services/ugc/research/mongodb/requirements.txt
pymongo[srv]==4.7.3
python-dotenv==1.0.1
